Separated the different spider classes and fixed the spider for USAToday to work properly
